# internal/domain/order.go
package models

import (
	"time"

	"github.com/google/uuid"
)

type Order struct {
	ID            uuid.UUID  `db:"id" json:"id"`
	UserID        *uuid.UUID `db:"user_id" json:"user_id,omitempty"`
	Status        string     `db:"status" json:"status"`
	TotalAmount   float64    `db:"total_amount" json:"total_amount"`
	PaymentMethod string     `db:"payment_method" json:"payment_method"`
	CreatedAt     time.Time  `db:"created_at" json:"created_at"`
	UpdatedAt     time.Time  `db:"updated_at" json:"updated_at"`
}

type OrderItem struct {
	ID        uuid.UUID `db:"id" json:"id"`
	OrderID   uuid.UUID `db:"order_id" json:"order_id"`
	ProductID uuid.UUID `db:"product_id" json:"product_id"`
	Quantity  int       `db:"quantity" json:"quantity"`
	Price     float64   `db:"price" json:"price"`
	CreatedAt time.Time `db:"created_at" json:"created_at"`
}

type OrderFilter struct {
	UserID *uuid.UUID
	Limit  int
	Offset int
}